// src/types/generated/AppConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Defaults } from "./Defaults";
import type { NotificationPrefs } from "./NotificationPrefs";
import type { ThemePref } from "./ThemePref";
import type { TunnelConfig } from "./TunnelConfig";

/**
 * The whole persisted application configuration.
 */
export type AppConfig = { version: number, theme: ThemePref, startOnBoot: boolean, minimizeToTray: boolean, notifications: NotificationPrefs, defaults: Defaults, tunnels: Array<TunnelConfig>, };